+++ src/components/radio/radio.ts
@@ -12,8 +12,6 @@ import {
   arrowRight,
   arrowUp,
 } from '../common/controllers/key-bindings.js';
-import { alternateName } from '../common/decorators/alternateName.js';
-import { blazorTwoWayBind } from '../common/decorators/blazorTwoWayBind.js';
 import { watch } from '../common/decorators/watch.js';
 import { registerComponent } from '../common/definitions/register.js';
 import messages from '../common/localization/validation-en.js';
@@ -133,12 +131,12 @@ export default class IgcRadioComponent extends FormAssociatedRequiredMixin(
     return this._value;
   }
 
+  /* @tsTwoWayProperty(true, "igcChange", "detail", false) */
   /**
    * The checked state of the control.
    * @attr
    */
   @property({ type: Boolean })
-  @blazorTwoWayBind('igcChange', 'detail')
   public set checked(value: boolean) {
     this._checked = Boolean(value);
     this._checked ? this._updateCheckedState() : this._updateUncheckedState();
@@ -188,14 +186,14 @@ export default class IgcRadioComponent extends FormAssociatedRequiredMixin(
     this.input.click();
   }
 
+  /* alternateName: focusComponent */
   /** Sets focus on the radio control. */
-  @alternateName('focusComponent')
   public override focus(options?: FocusOptions) {
     this.input.focus(options);
   }
 
+  /* alternateName: blurComponent */
   /** Removes focus from the radio control. */
-  @alternateName('blurComponent')
   public override blur() {
     this.input.blur();
   }